threading for URL generation, Session() removed
